+++ src/crypto/iteratedhash.hpp
@@ -127,7 +127,7 @@ size_t iterated_hash<T>::hash(const char * input, size_t length) {
 		do {
 			
 			hash_word buffer[block_size / sizeof(hash_word)];
-			byte_order::load(input, buffer, boost::size(buffer));
+			byte_order::load(input, buffer, size_t(boost::size(buffer)));
 			
 			transform::transform(state, buffer);
